requirements.txt
   websockets==14.1
   pytest==8.3.4
   pytest-asyncio==0.25.1
   allure-pytest==2.13.5
   numpy==1.26.4
